requirements.txt
django>=1.5.4
django-model-utils>=1.5.0
git+git://github.com/maraujop/django-crispy-forms.git@dev#egg=django-crispy-forms
django-allauth
pillow
markdown
django-vanilla-views
django-permanent
south